// types.ts
export interface DistributionData {
    name: string;
    value: number;
    students: number;
    color: string;
}

export interface GenderComparisonData {
    name: string;
    femenino: number;
    masculino: number;
}

export interface GradeCompositionData {
    name: string;
    delgadez: number;
    riesgoDelgadez: number;
    adecuado: number;
    sobrepeso: number;
    obesidad: number;
}